+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - [vtex login, vtex switch] Prompt workspace creation if it doesn't exist.
 
+### Added
+- [vtex link] Generate `linkID` and set the `x-vtex-bh-link-id` header on link and relink requests to `builder-hub`.
+
 ## [2.101.0] - 2020-05-19
 ### Added
 - [vtex publish] Message referencing location of the published docs if any.